✨ Refactor New Project Page
- Rename NewProjectPage to ProjectNewPage
- Update import and export paths
- Add Form component for creating new projects
- Create corresponding CSS module for styling
