--- frontend/apps/app/features/projects/pages/NewProjectPage/index.ts
@@ -1,1 +1,0 @@
-export * from './NewProjectPage'
