+++ frontend/apps/app/features/projects/pages/ProjectNewPage/ProjectNewPage.tsx
@@ -1,8 +1,8 @@
 import type React from 'react'
 import { Form } from './Form'
-import styles from './NewProjectPage.module.css'
+import styles from './ProjectNewPage.module.css'
 
-export const NewProjectPage: React.FC = () => {
+export const ProjectNewPage: React.FC = () => {
   return (
     <div className={styles.container}>
       <Form />
