+++ frontend/apps/app/features/projects/pages/ProjectNewPage/index.ts
@@ -1,0 +1,1 @@
+export * from './ProjectNewPage'
