+++ frontend/apps/app/features/projects/pages/index.ts
@@ -1,3 +1,3 @@
 export * from './ProjectDetailPage'
 export * from './ProjectsPage'
-export * from './NewProjectPage'
+export * from './ProjectNewPage'
